requirements.txt
ultralytics>=8.2.0
roboflow>=1.1.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
pandas>=2.0.0
PyYAML>=6.0
torch>=2.0.0
torchvision>=0.15.0
